+++ backend/src/main/java/com/skapp/community/leaveplanner/constant/LeaveMessageConstant.java
@@ -128,7 +128,6 @@ public enum LeaveMessageConstant implements MessageConstant {
 			"api.error.leave.policy-leave-invalid-status-transition-employee"),
 	LEAVE_ERROR_POLICY_LEAVE_ALREADY_IN_STATUS("api.error.leave.policy-leave-already-in-status"),
 	LEAVE_ERROR_POLICY_LEAVE_INFORMANT_CANNOT_REVIEW("api.error.leave.policy-leave-informant-cannot-review"),
-	LEAVE_ERROR_POLICY_LEAVE_DECLINE_REASON_REQUIRED("api.error.leave.policy-leave-decline-reason-required"),
 	LEAVE_ERROR_POLICY_LEAVE_REVIEWER_COMMENT_MAX_LENGTH("api.error.leave.policy-leave-reviewer-comment-max-length"),
 	LEAVE_ERROR_POLICY_LEAVE_NUDGE_TOO_SOON("api.error.leave.policy-leave-nudge-too-soon"),
 	LEAVE_ERROR_POLICY_LEAVE_INVALID_PAGE("api.error.leave.policy-leave-invalid-page"),
