--- backend/src/main/java/com/skapp/community/leaveplanner/payload/request/PolicyLeaveCancelRequestDto.java
@@ -1,15 +1,0 @@
-package com.skapp.community.leaveplanner.payload.request;
-
-import com.skapp.community.leaveplanner.type.LeaveRequestStatus;
-import jakarta.validation.constraints.NotNull;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class PolicyLeaveCancelRequestDto {
-
-	@NotNull
-	private LeaveRequestStatus status;
-
-}
