+++ backend/src/main/java/com/skapp/community/leaveplanner/repository/PolicyLeaveRequestDao.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PolicyLeaveRequestDao
-		extends JpaRepository<PolicyLeaveRequest, Long>, PolicyLeaveRequestRepository, PolicyLeaveReviewRepository {
+public interface PolicyLeaveRequestDao extends JpaRepository<PolicyLeaveRequest, Long>, PolicyLeaveRequestRepository {
 
 }
