+++ backend/src/main/java/com/skapp/community/leaveplanner/repository/PolicyLeaveRequestRepository.java
@@ -9,11 +9,19 @@ import org.springframework.data.domain.Pageable;
 import java.time.LocalDate;
 import java.util.Collection;
 import java.util.List;
+import java.util.Optional;
 
 public interface PolicyLeaveRequestRepository {
 
 	Page<PolicyLeaveRequest> findMyRequests(Long employeeId, LocalDate cycleStart, LocalDate cycleEnd,
 			PolicyLeaveRequestFilterDto filterDto, Pageable pageable);
+
+	Page<PolicyLeaveRequest> findSupervisedRequests(Long supervisorEmployeeId,
+			PolicyLeaveRequestFilterDto filterDto, Pageable pageable);
+
+	List<PolicyLeaveRequest> findPendingSupervisedRequests(Long supervisorEmployeeId, String searchKeyword);
+
+	Optional<PolicyLeaveRequest> findByIdForUpdate(Long id);
 
 	Double sumCommittedDaysForPolicyInCycle(Long employeeId, Long policyId, Collection<LeaveRequestStatus> statuses,
 			LocalDate cycleStart, LocalDate cycleEnd);
