+++ backend/src/main/java/com/skapp/community/leaveplanner/service/PolicyLeaveService.java
@@ -4,6 +4,7 @@ import com.skapp.community.common.payload.response.ResponseEntityDto;
 import com.skapp.community.leaveplanner.payload.request.PolicyLeaveAvailabilityRequestDto;
 import com.skapp.community.leaveplanner.payload.request.PolicyLeaveRequestDto;
 import com.skapp.community.leaveplanner.payload.request.PolicyLeaveRequestFilterDto;
+import com.skapp.community.leaveplanner.payload.request.PolicyLeaveRequestUpdateDto;
 
 public interface PolicyLeaveService {
 
@@ -14,5 +15,17 @@ public interface PolicyLeaveService {
 	ResponseEntityDto applyPolicyLeaveRequest(PolicyLeaveRequestDto policyLeaveRequestDto);
 
 	ResponseEntityDto getCurrentUserPolicyLeaveRequests(PolicyLeaveRequestFilterDto policyLeaveRequestFilterDto);
+
+	ResponseEntityDto getSupervisedPolicyLeaveRequests(PolicyLeaveRequestFilterDto policyLeaveRequestFilterDto);
+
+	ResponseEntityDto getPendingSupervisedPolicyLeaveRequests(String searchKeyword);
+
+	ResponseEntityDto getPolicyLeaveRequestById(Long id);
+
+	ResponseEntityDto updatePolicyLeaveRequest(Long id, PolicyLeaveRequestUpdateDto policyLeaveRequestUpdateDto);
+
+	ResponseEntityDto nudgePolicyLeaveRequestManagers(Long id);
+
+	ResponseEntityDto getPolicyLeaveRequestNudgeStatus(Long id);
 
 }
