+++ backend/src/main/resources/community/messages/leave-messages.properties
@@ -109,7 +109,6 @@ api.error.leave.policy-leave-invalid-status-transition-manager=This leave reques
 api.error.leave.policy-leave-invalid-status-transition-employee=This leave request can no longer be cancelled
 api.error.leave.policy-leave-already-in-status=This leave request is already in the selected status
 api.error.leave.policy-leave-informant-cannot-review=Informants cannot review leave requests
-api.error.leave.policy-leave-decline-reason-required=A reason is required to decline a leave request
 api.error.leave.policy-leave-reviewer-comment-max-length=The reason cannot exceed 255 characters
 api.error.leave.policy-leave-nudge-too-soon=This leave request has already been nudged within the last 24 hours
 api.error.leave.policy-leave-invalid-page=The page number cannot be negative
